update colors, inputs, buttons, created move up botton
