+++ index.html
@@ -24,18 +24,19 @@
     <nav>
       <div class="country-list">
         <div class="logo">
-          <a href="#">
+          <a href="./index.html">
             <img src="./icons/bet.png" author="freepik" alt="logo ball on books">
             
           </a>
           <h1>Football Scores</h1>
+          
         </div>
           
         
         <p class="p-active">Poland</p>
         <p>England</p>
         <form>
-          <input type="text" id="country-input">
+          <input type="text" id="country-input" placeholder="type country">
           <button type="button" class="country-button">Check</button>
         </form>
         <p>Germany</p>
@@ -75,10 +76,8 @@
           <button type="submit">Wyślij mnie</button>
       </form>
       </section> -->
-    
-    <script type="module" src="./render.js"></script>
-    <script  type="module" src="./additions.js"></script>
-    <script  type="module" src="main.js"></script>
+    <div class="up-button"><img src="./icons/up-arrow.png" alt="arrow up"></div>
+    <script  type="module" src="./js/main.js"></script>
   </body>
 </html>
