+++ js/env.js
@@ -1,0 +1,9 @@
+const options = {
+    method: "GET",
+    headers: {
+      "X-RapidAPI-Key": "dc935f1d6emsh3dde368f85b528ep1aca42jsnaaeaf9657fab",
+      "X-RapidAPI-Host": "api-football-v1.p.rapidapi.com",
+    },
+  };
+
+  export {options}
