+++ js/render.js
@@ -1,7 +1,124 @@
+export { createLiveMatches, generateEventsMatch };
+import { getNormalDate } from "./additions.js";
+import { createListenerMatch } from "./matches.js";
 
-// index
+function generateEventHTML(event) {
+  if (event === "Yellow Card") {
+    return `<img src="./icons/matches_icon/yellow-card.png" alt="yellow card">`;
+  }
+  if (event === "Normal Goal" || event === "Own Goal") {
+    return `<img src="./icons/matches_icon/goal.png" alt="football ball">`;
+  }
+  if (event === "Red Card") {
+    return `<img src="./icons/matches_icon/red.png" alt="red card">`;
+  } else {
+    return `<p class="event-text">${event}</p>`;
+  }
+}
+
+function generateEventsMatch(table, el) {
+  const teamLeft = el.querySelector(".team-left");
+  const teamRight = el.querySelector(".team-right");
+  let homeAll = "";
+  let awayAll = "";
+
+  teamLeft.innerHTML = "";
+  teamRight.innerHTML = "";
+
+  for (let i = 0; i < table.length; i++) {
+    let home = "";
+    let away = "";
+    let event = generateEventHTML(table[i].events.detail);
+    if (table[i].events.type !== "subst") {
+      if (
+        table[i].events.team.name ===
+        el.querySelector(".team-name-first").innerText
+      ) {
+        home = `<div class="event-left"><p class="event-time">${table[i].events.time.elapsed}'</p><p class="event-player">${table[i].events.player.name}</p><p class="event-detail">${event}</p>
+      </div>`;
+        homeAll = homeAll + home;
+      }
+      if (
+        table[i].events.team.name ===
+        el.querySelector(".team-name-second").innerText
+      ) {
+        away = `<div class="event-right"><p class="event-time">${table[i].events.time.elapsed}'</p><p class="event-player">${table[i].events.player.name}</p><p class="event-detail">${event}</p></div>
+      `;
+        awayAll = awayAll + away;
+      }
+    }
+  }
+  teamRight.innerHTML = awayAll;
+  teamLeft.innerHTML = homeAll;
+}
+
+function createLiveMatches(table) {
+  const detailsTable = [];
+  const body = document.querySelector(".matches-body");
+  body.innerHTML = "";
+  let patternAll = "";
+  let isWinner1 = "";
+  let isWinner2 = "";
+  //  console.log(table)
+  for (let i = 0; i < table.length; i++) {
+    if (
+      table[i].matchInfo.matchTime === null ||
+      table[i].score.home === null ||
+      table[i].score.away === null
+    ) {
+      table[i].matchInfo.matchTime = table[i].matchInfo.status;
+      table[i].score.home = "-";
+      table[i].score.away = "-";
+    }
+    if (table[i].matchInfo.status.toLowerCase() === "match finished") {
+      if (table[i].score.home === table[i].score.away) {
+        //
+      } else if (table[i].teams.home.winner === true) {
+        isWinner1 = `winner`;
+      } else {
+        isWinner2 = `winner`;
+      }
+    }
+
+    let pattern = ` <div data-id=${i} data-fixture=${
+      table[i].matchInfo.fixtureId
+    } class="match">
+        <p>${getNormalDate(table[i].matchInfo.data)}</p>
+        <div class="match-top">
+        <p class="team-name-first ${isWinner1}">${table[i].teams.home.name}</p>
+        <p class="team-name-second ${isWinner2}">${table[i].teams.away.name}</p>
+        <div class="team-1">
+                <img src=${table[i].teams.home.logo} alt=" logo${
+      table[i].teams.away.name
+    }">
+            </div>
+            <p class="time">${table[i].matchInfo.matchTime}'</p>  
+            <div class="score-1">
+            <span>${table[i].score.home}</span>
+             </div>
+             <div class="score-2">
+             <span>${table[i].score.away}</span>
+             </div>
+            <div class="team-2">
+                <img src=${table[i].teams.away.logo} alt="logo ${
+      table[i].teams.away.name
+    }">
+            </div>
+        </div>
+        <div class="match-bottom">
+            <div class="team-left"></div>
+            <div class="team-right"></div>
+            </div>
+       </div>`;
+
+    patternAll += pattern;
+    isWinner1 = "";
+    isWinner2 = "";
+  }
+  body.innerHTML = patternAll;
+  createListenerMatch(table, detailsTable);
+}
 const createCountryElement = (countries) => {
-  console.log(countries);
   const body = document.querySelector(".league-body");
   const leagues = document.querySelector(".table");
   const players = document.querySelector(".players");
@@ -11,18 +128,12 @@ const createCountryElement = (countries) => {
   let patternAll = "";
 
   for (let i = 0; i < countries.length; i++) {
-    let pattern = ` <div data-id=${countries[i].league.id} class="league-box">
-		
-		 <div class="league-box-mid">
-     
+    let pattern = ` <div data-id=${countries[i].league.id} class="league-box">	
+		 <div class="league-box-mid">   
 		   <div class="league-box-mid-img">
 			 <img src=${countries[i].league.png} alt=${countries[i].league.name} >
 		   <p>${countries[i].league.name}</p>
-		   </div>
-		   
-			 
-		   
-		   
+		   </div>	   
 		 </div>
 	 </div>`;
 
@@ -32,25 +143,22 @@ const createCountryElement = (countries) => {
   body.innerHTML = patternAll;
 };
 function renderContent(callback, selectorCss) {
+
   let choose;
-  let leagues = document.querySelectorAll(selectorCss);
-  console.log(leagues);
+  const leagues = document.querySelectorAll(selectorCss);
   leagues.forEach((el) => {
     el.addEventListener("click", function () {
-      console.log(el.getAttribute("data-id") + " z listenera");
       choose = el.getAttribute("data-id");
       callback(choose);
     });
   });
 }
 function createLeagueElement(leagues) {
-  console.log(leagues);
   const body = document.querySelector(".table");
   body.innerHTML = "";
   let patternAll = "";
   let stats = `<div class="table-box"><div class="table-box-text">
   <span class="lp">#</span>
-
   <span class="name">Team</span>
   <span class="matches">Matches</span>
   <span class="wins">Wins</span>
@@ -58,18 +166,16 @@ function createLeagueElement(leagues) {
   <span class="lose">Loses</span>
   <span class="goals">Balance</span>
   <span class="points">Points</span>
-  <span class="balance">Last matches</span>
-  
+  <span class="balance">Last matches</span>  
 </div></div>`;
 
   if (leagues === false) {
-    let errorMessage = `<div class="error"><p>We can't download this data, sorry</p></div>`;
+    const errorMessage = `<div class="error"><p>We can't download this data, sorry</p></div>`;
     body.innerHTML = errorMessage;
   } else {
     for (let i = 0; i < leagues.length; i++) {
-      let form = generateFormHTML(leagues[i].form)
+      let form = generateFormHTML(leagues[i].form);
       let pattern = `<div data-id=${leagues[i].team_id} class="table-box">
-  
       <div class="table-box-text">
         <span class="lp">${leagues[i].rank}</span>
         <img src=${leagues[i].logo} alt=${leagues[i].team}>
@@ -84,40 +190,33 @@ function createLeagueElement(leagues) {
           ${form}
         </div>
       </div>
-  
     </div>`;
 
       patternAll = patternAll + pattern;
     }
 
     body.innerHTML = stats + patternAll;
-    body.scrollIntoView({behavior:"smooth"})
+    body.scrollIntoView({ behavior: "smooth" });
   }
-  
 }
 
-function generateFormHTML(formString){
-  
-
-  let arr = formString.split('');
-  for(let i =0;i<arr.length;i++){
-    if(arr[i] === "W"){
-      arr[i] = `<span class="form-win">${arr[i]}</span>`
+function generateFormHTML(formString) {
+  const arr = formString.split("");
+  for (let i = 0; i < arr.length; i++) {
+    if (arr[i] === "W") {
+      arr[i] = `<span class="form-win">${arr[i]}</span>`;
     }
-    if(arr[i] === "D"){
-      arr[i] =`<span class="form-remis">${arr[i]}</span>`
+    if (arr[i] === "D") {
+      arr[i] = `<span class="form-remis">${arr[i]}</span>`;
     }
-    if(arr[i] === "L"){
-      arr[i] = `<span class="form-lose">${arr[i]}</span>`
+    if (arr[i] === "L") {
+      arr[i] = `<span class="form-lose">${arr[i]}</span>`;
     }
   }
-    return arr.join('');
-  
-  
+  return arr.join("");
 }
 
 function createElementPlayers(players) {
-  console.log(players);
   const body = document.querySelector(".players");
 
   let patternAll = "";
@@ -126,7 +225,7 @@ function createElementPlayers(players) {
   playersBox.classList.add("players-box");
   const logoBox = document.createElement("div");
   logoBox.classList.add("players-logo-box");
-  let logoSrc = [];
+  const logoSrc = [];
   const teamLogo = document.querySelectorAll(".table-box");
   teamLogo.forEach((el) => {
     if (el.getAttribute("data-id") === players[0].team_id) {
@@ -135,22 +234,20 @@ function createElementPlayers(players) {
       // logoSrc.push(el.querySelector(".table-box-text name").innerText);
     }
   });
-  let header =`<div class="player-box top">
+  const header = `<div class="player-box top">
   <p class="player-name"><img src="./icons/id.png" alt=""> Name</p>
      <p class="player-age"> <img src="./icons/age.png" alt=""> Age</p>
      <p class="player-number"><img src="./icons/shirt.png" alt=""> Number</p>
      <p class="player-position"><img src="./icons/location.png" alt=""> Position</p>
    </div>`;
-  let logo = `<div class="players-logo-box-text">
+  const logo = `<div class="players-logo-box-text">
   <h3>${logoSrc[1]}</h3>
   <img src="${logoSrc[0]}">
-  
 </div>
-
 <h4>Squad</h4>`;
 
   for (let i = 0; i < players.length; i++) {
-    let positionIcon = generatePlayerPositionHTML(players[i].position)
+    let positionIcon = generatePlayerPositionHTML(players[i].position);
     let pattern = `
     <div class="player-box" data-id=${players[i].player_id}>
     <div class="photos"><img class="first"src=${logoSrc[0]} alt=${logoSrc[1]}>
@@ -167,28 +264,27 @@ function createElementPlayers(players) {
   playersBox.innerHTML = header + patternAll;
   body.appendChild(logoBox);
   body.appendChild(playersBox);
-  body.scrollIntoView({behavior:"smooth"})
+  body.scrollIntoView({ behavior: "smooth" });
 }
-function generatePlayerPositionHTML(playerPos){
-  if(playerPos ==="Goalkeeper"){
-    return `<img src="./icons/players_icon/goalkeeper.png" author="Freepik" alt="goalkeeper icon">`
+function generatePlayerPositionHTML(playerPos) {
+  if (playerPos === "Goalkeeper") {
+    return `<img src="./icons/players_icon/goalkeeper.png" author="Freepik" alt="goalkeeper icon">`;
   }
-  if(playerPos ==="Defender"){
+  if (playerPos === "Defender") {
     return `<img src="./icons/players_icon/shield.png"
-    author="Freepik" alt="shield icon">`
+    author="Freepik" alt="shield icon">`;
   }
-  if(playerPos ==="Midfielder"){
+  if (playerPos === "Midfielder") {
     return `<img src="./icons/players_icon/injury.png"
-    author="Freepik" alt="shirt with plus icon">`
+    author="Freepik" alt="shirt with plus icon">`;
   }
-  if(playerPos ==="Attacker"){
+  if (playerPos === "Attacker") {
     return `<img src="./icons/players_icon/soccer-player.png"
-    author="Freepik" alt="soccer player icon">`
+    author="Freepik" alt="soccer player icon">`;
   }
 }
 
 function createPlayerStatistics(players) {
-  console.log(players);
   const body = document.querySelector(".player-stats");
   body.innerHTML = "";
   let documentBody = "";
@@ -207,7 +303,9 @@ function createPlayerStatistics(players) {
   </div>`;
 
   for (let i = 0; i < players.statistics.length; i++) {
-    let position = generatePlayerPositionHTML(players.statistics[i].statistics.games.position)
+    let position = generatePlayerPositionHTML(
+      players.statistics[i].statistics.games.position
+    );
     let tournamentPattern = `<div class="player-stats-turnaments">
     <p class="season">season ${players.statistics[i].statistics.league.season}</p>
     <div class="turnaments-top">
@@ -245,22 +343,6 @@ function createPlayerStatistics(players) {
      <span>accuracy:</span><p class="accuracy"> ${players.statistics[i].statistics.passes.accuracy}</p> </div>
      <div class="turnaments-box"><h5><img src="./icons/stats_icon/bullet.png" author="Smashicons" alt="bullet in fire">Shots</h5> <span>total shots:</span><p class="total"> ${players.statistics[i].statistics.shots.total}</p>
      <span>On enemy gate:</span><p class="on"> ${players.statistics[i].statistics.shots.on}</p></div>
-     
-     
-     
-     
-     
- 
-     
-   
-     
-    
-     
-    
-     
-     
-     
-     
   </div>
   </div>`;
 
@@ -269,10 +351,8 @@ function createPlayerStatistics(players) {
 
   documentBody = patternInfo + tournamentInfo;
   body.innerHTML = documentBody;
-  body.scrollIntoView({behavior:"smooth"})
+  body.scrollIntoView({ behavior: "smooth" });
 }
-
-// matches
 
 function choosePlayer(callback) {
   const body = document.querySelectorAll(".player-box");
