+++ matches.html
@@ -17,7 +17,10 @@
 <body>
     <div class="container">
             <header>
+                <a href="./index.html">
                 <img src="./icons/bet.png" author="freepik" alt="logo ball on books">
+                    
+                </a>
                 <h1>Matches</h1>
             </header>
         
@@ -62,9 +65,7 @@
         </div>
        
        </div> 
-    <script type="module" src="./render.js"></script>
-    <script type="module" src="./additions.js"></script>
-    <script type="module" src="./matches.js"></script>
+    <script type="module" src="./js/matches.js"></script>
 
 
 </body>
