+++ scss/_button.scss
@@ -1,0 +1,79 @@
+.country-button{
+  
+    margin: .5em;
+    align-items: center;
+    background-color: $main-color;
+    border: 3px solid $font-color;
+    border-radius: 15px;
+    box-sizing: border-box;
+    color: $font-color;
+    cursor: pointer;
+    display: inline-flex;
+    fill: #000;
+    // font-size: 16px;
+    font-weight: 600;
+    // height: 48px;
+    justify-content: center;
+    letter-spacing: -.8px;
+    line-height: 24px;
+    // min-width: 140px;
+    outline: 0;
+    padding: 0 17px;
+    text-align: center;
+    text-decoration: none;
+    transition: all .3s;
+    user-select: none;
+    -webkit-user-select: none;
+    touch-action: manipulation;
+    width: 100%;
+    a{
+      text-decoration: none;
+      color: $font-color;
+    }
+  }
+
+  .country-button:hover {
+    background-color: $font-color;
+    color: $bg-color;
+  
+  }
+
+  //matches
+.form-submit,.back{
+    margin: 1em;
+    align-items: center;
+    background-color: $main-color;
+    border: 3px solid $font-color;
+    box-sizing: border-box;
+    color: $font-color;
+    cursor: pointer;
+    display: inline-flex;
+    fill: #000;
+    font-size: 16px;
+    font-weight: 600;
+    height: 48px;
+    justify-content: center;
+    letter-spacing: -.8px;
+    line-height: 24px;
+    min-width: 140px;
+    outline: 0;
+    padding: 0 17px;
+    text-align: center;
+    text-decoration: none;
+    transition: all .3s;
+    user-select: none;
+    -webkit-user-select: none;
+    touch-action: manipulation;
+    a{
+      text-decoration: none;
+      color: $font-color;
+    }
+  }
+
+  
+  .form-submit:hover,.back:focus {
+    background-color: $font-color;
+    color: $bg-color;
+  
+  }
+
