+++ scss/_colors.scss
@@ -1,4 +1,4 @@
 $bg-color: rgb(41, 41, 41);
 $main-color: rgb(57, 57, 57);
-$font-color: rgb(116, 117, 118);
+$font-color: rgb(182, 183, 185);
 $active-color: rgb(191, 158, 241);
