+++ scss/_reset.scss
@@ -1,0 +1,8 @@
+
+*,
+*::after,
+*::before {
+  box-sizing: border-box;
+  padding: 0;
+  margin: 0;
+}
