+++ scss/main.scss
@@ -1,13 +1,7 @@
 @import "colors";
+@import "reset";
+@import "button";
 
-
-*,
-*::after,
-*::before {
-  box-sizing: border-box;
-  padding: 0;
-  margin: 0;
-}
 // liga player tablea
 html {
   font-family: "Roboto", sans-serif;
@@ -72,64 +66,32 @@ form{
   align-items: center;
   input{
     background-color: $font-color;
-    max-width: 150px;
+    max-width: 200px;
+   
   color: #ffffff;
   font-size: 18px;
   border: none;
   outline: none;
-  border-radius: 5px;
+  border-radius: 15px;
   padding: .3em;
   }
 }
+.up-button{
+  position: fixed;
 
-.country-button{
-  
-  margin: .5em;
-  align-items: center;
-  background-color: $main-color;
-  border: 3px solid $font-color;
-  box-sizing: border-box;
-  color: $font-color;
+  background-color: $font-color;
+  bottom: 0;
+  left: 80%;
+  border-top-left-radius: 15px;
+  border-top-right-radius: 15px;
+  padding: .3em;
+  transform: translateX(-50%,-50%);
   cursor: pointer;
-  display: inline-flex;
-  fill: #000;
-  // font-size: 16px;
-  font-weight: 600;
-  // height: 48px;
-  justify-content: center;
-  letter-spacing: -.8px;
-  line-height: 24px;
-  // min-width: 140px;
-  outline: 0;
-  padding: 0 17px;
-  text-align: center;
-  text-decoration: none;
-  transition: all .3s;
-  user-select: none;
-  -webkit-user-select: none;
-  touch-action: manipulation;
-  width: 100%;
-  a{
-    text-decoration: none;
-    color: $font-color;
+  img{
+    width: 30px;
   }
 }
 
-  .country-button:focus {
-  color: #171e29;
-}
-
-.country-button:hover {
-  border-color: #06f;
-  color: #06f;
-  fill: #06f;
-}
-
-.country-button:active {
-  border-color: #06f;
-  color: $font-color;
-  fill: #06f;
-}
 
 
 
@@ -159,15 +121,17 @@ form{
     display: block;
     color: $active-color;
   }
-}
-
-.p-active {
-  position: relative;
-  border: 2px solid $active-color;
-  p {
-    color: $active-color;
+  .p-active {
+    background-color: $font-color;
+    color: $bg-color;
+  
+  }
+  .p-active:hover{
+    color: $bg-color;
   }
 }
+
+
 
 .countries {
   display: flex;
@@ -539,6 +503,7 @@ form{
           grid-column: 1/-1;
           text-align: center;
           margin-bottom: .5em;
+          font-size: 1.2rem;
           img{
             width: 20px;
             margin-right: .2em;
