+++ scss/matches.scss
@@ -1,5 +1,7 @@
 @import "colors";
 
+@import "button";
+
 .container{
   margin: 0 auto;
   max-width: 1200px;
@@ -47,58 +49,8 @@ form{
 
 }
 
-button{
-  margin: 1em;
-  align-items: center;
-  background-color: $main-color;
-  border: 3px solid $font-color;
-  box-sizing: border-box;
-  color: $font-color;
-  cursor: pointer;
-  display: inline-flex;
-  fill: #000;
-  font-size: 16px;
-  font-weight: 600;
-  height: 48px;
-  justify-content: center;
-  letter-spacing: -.8px;
-  line-height: 24px;
-  min-width: 140px;
-  outline: 0;
-  padding: 0 17px;
-  text-align: center;
-  text-decoration: none;
-  transition: all .3s;
-  user-select: none;
-  -webkit-user-select: none;
-  touch-action: manipulation;
-  a{
-    text-decoration: none;
-    color: $font-color;
-  }
-}
 
-button:focus {
-  color: #171e29;
-}
 
-button:hover {
-  border-color: #06f;
-  color: #06f;
-  fill: #06f;
-}
-
-button:active {
-  border-color: #06f;
-  color: $font-color;
-  fill: #06f;
-}
-
-@media (min-width: 768px) {
-  .button-59 {
-    min-width: 170px;
-  }
-}
 .date-to,.date-from,select{
   background-color: $font-color;
   font-family: "Roboto Mono",monospace;
@@ -286,4 +238,8 @@ button:active {
   }
 
 
-
+  @media (min-width: 768px) {
+    .button-59 {
+      min-width: 170px;
+    }
+  }
