--- testowy.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Document</title>
-    <link rel="stylesheet" href="./css/main.css">
-</head>
-<body>
-
-        <section id="form-contact">
-            <form action="">
-                <label for="email">E-mail:</label>
-                <input type="text" id="email" class="email">
-                <label for="message">Message:</label>
-                <input type="text-area" id="message" class="message">
-                <button>Submit</button>
-
-            </form>
-
-        </section>
-
-
-    <script src="./main.js"></script>
-</body>
-</html>
